Working update example
I've removed all the references to &mut that I added recently. They made
the library difficult to use. This means the library is NOT thread-safe,
although I guess that was the case anyway. If used in a single-threaded
context, I don't believe it is possible to leave in an inconsistent
state, except by premature exit (SIGINT and friends)